Add syntax highlighting for js and react
These plugins make the colors look nicer when developing javascript and
react code.

I tried to enable vim-styled-jsx as well, but ran into an issue with
load order and couldn't get it to work.

https://github.com/alampros/vim-styled-jsx/issues/1

I tried renaming the plugin folder A_vim-styled-jsx to fix loading order
(as suggested here:
https://stackoverflow.com/questions/20036921/vim-pathogen-and-load-order-of-ftplugin-files)
but that didn't work.